rename binary tree pageand delete old binary tree page
