--- visualization/algovis/src/app/binarySearch2/layout.tsx
@@ -1,11 +1,0 @@
-export default function RootLayout({
-    children,
-  }: {
-    children: React.ReactNode
-  }) {
-    return (
-      <html lang="en">
-        <body>{children}</body>
-      </html>
-    )
-  }
